refactor/fix tests #1
