+++ jobs/jobs_test.go
@@ -1,6 +1,7 @@
 package jobs
 
 import (
+	"os"
 	"testing"
 
 	"github.com/gin-gonic/gin"
@@ -12,4 +13,5 @@ func TestMain(m *testing.M) {
 	utils.SetTesting("../.env")
 	models.Connect(utils.Env.TestDatabaseURL)
 	gin.SetMode(gin.TestMode)
+	os.Exit(m.Run())
 }
