+++ models/models_test.go
@@ -1,0 +1,11 @@
+package models
+
+import (
+	"os"
+	"testing"
+)
+
+func TestMain(m *testing.M) {
+	setupModelsTests()
+	os.Exit(m.Run())
+}
