+++ models/models_test_utils.go
@@ -14,11 +14,6 @@ import (
 	"github.com/teris-io/shortid"
 )
 
-func TestMain(m *testing.M) {
-	setupModelsTests()
-	os.Exit(m.Run())
-}
-
 func setupModelsTests() {
 	utils.SetTesting("../.env")
 	Connect(utils.Env.TestDatabaseURL)
