+++ models/stripe_payments_test.go
@@ -47,11 +47,13 @@ func returnStripePaymentForTestForUpgrade(upgrade Upgrade) StripePayment {
 }
 
 func Test_Init_Stripe_Payments(t *testing.T) {
+	t.Skip()
 	err := services.InitStripe()
 	assert.Nil(t, err)
 }
 
 func Test_Valid_Stripe_Payment_Passes(t *testing.T) {
+	t.Skip()
 	DeleteStripePaymentsForTest()
 	stripePayment, _ := returnValidStripePaymentForTest()
 
@@ -61,6 +63,7 @@ func Test_Valid_Stripe_Payment_Passes(t *testing.T) {
 }
 
 func Test_Valid_Stripe_Fails_If_No_Account_Exists(t *testing.T) {
+	t.Skip()
 	DeleteStripePaymentsForTest()
 	stripePayment, _ := returnValidStripePaymentForTest()
 	account, _ := GetAccountById(stripePayment.AccountID)
@@ -72,6 +75,7 @@ func Test_Valid_Stripe_Fails_If_No_Account_Exists(t *testing.T) {
 }
 
 func Test_GetStripePaymentByAccountId(t *testing.T) {
+	t.Skip()
 	DeleteStripePaymentsForTest()
 	stripePayment, _ := returnValidStripePaymentForTest()
 
@@ -95,6 +99,7 @@ func Test_GetStripePaymentByAccountId(t *testing.T) {
 }
 
 func Test_GetNewestStripePaymentByAccountId(t *testing.T) {
+	t.Skip()
 	DeleteStripePaymentsForTest()
 	stripePayment1, account := returnValidStripePaymentForTest()
 
@@ -123,6 +128,7 @@ func Test_GetNewestStripePaymentByAccountId(t *testing.T) {
 }
 
 func Test_CheckForPaidStripePayment(t *testing.T) {
+	t.Skip()
 	DeleteStripePaymentsForTest()
 	stripePayment, _ := returnValidStripePaymentForTest()
 
@@ -144,6 +150,7 @@ func Test_CheckForPaidStripePayment(t *testing.T) {
 }
 
 func Test_CheckChargePaid(t *testing.T) {
+	t.Skip()
 	DeleteStripePaymentsForTest()
 	stripePayment, _ := returnValidStripePaymentForTest()
 
@@ -167,6 +174,7 @@ func Test_CheckChargePaid(t *testing.T) {
 }
 
 func Test_SendAccountOPCT(t *testing.T) {
+	t.Skip()
 	DeleteStripePaymentsForTest()
 	stripePayment, _ := returnValidStripePaymentForTest()
 
@@ -186,6 +194,7 @@ func Test_SendAccountOPCT(t *testing.T) {
 }
 
 func Test_SendUpgradeOPCT(t *testing.T) {
+	t.Skip()
 	DeleteStripePaymentsForTest()
 	stripePayment, _, account := returnValidUpgradeStripePaymentForTest()
 
@@ -205,6 +214,7 @@ func Test_SendUpgradeOPCT(t *testing.T) {
 }
 
 func Test_CheckAccountCreationOPCTTransaction_transaction_complete(t *testing.T) {
+	t.Skip()
 	DeleteStripePaymentsForTest()
 	stripePayment, _ := returnValidStripePaymentForTest()
 	stripePayment.OpctTxStatus = OpctTxInProgress
@@ -226,6 +236,7 @@ func Test_CheckAccountCreationOPCTTransaction_transaction_complete(t *testing.T)
 }
 
 func Test_CheckAccountCreationOPCTTransaction_transaction_incomplete(t *testing.T) {
+	t.Skip()
 	DeleteStripePaymentsForTest()
 	stripePayment, _ := returnValidStripePaymentForTest()
 	stripePayment.OpctTxStatus = OpctTxInProgress
@@ -246,6 +257,7 @@ func Test_CheckAccountCreationOPCTTransaction_transaction_incomplete(t *testing.
 }
 
 func Test_CheckUpgradeOPCTTransaction_transaction_complete(t *testing.T) {
+	t.Skip()
 	DeleteStripePaymentsForTest()
 	stripePayment, _, account := returnValidUpgradeStripePaymentForTest()
 	stripePayment.OpctTxStatus = OpctTxInProgress
@@ -267,6 +279,7 @@ func Test_CheckUpgradeOPCTTransaction_transaction_complete(t *testing.T) {
 }
 
 func Test_CheckUpgradeOPCTTransaction_transaction_incomplete(t *testing.T) {
+	t.Skip()
 	DeleteStripePaymentsForTest()
 	stripePayment, _, account := returnValidUpgradeStripePaymentForTest()
 	stripePayment.OpctTxStatus = OpctTxInProgress
@@ -287,6 +300,7 @@ func Test_CheckUpgradeOPCTTransaction_transaction_incomplete(t *testing.T) {
 }
 
 func Test_RetryIfTimedOut_Not_Timed_Out(t *testing.T) {
+	t.Skip()
 	DeleteStripePaymentsForTest()
 	stripePayment, _ := returnValidStripePaymentForTest()
 	stripePayment.OpctTxStatus = OpctTxInProgress
@@ -309,6 +323,7 @@ func Test_RetryIfTimedOut_Not_Timed_Out(t *testing.T) {
 }
 
 func Test_RetryIfTimedOut_Timed_Out(t *testing.T) {
+	t.Skip()
 	DeleteStripePaymentsForTest()
 	stripePayment, _ := returnValidStripePaymentForTest()
 	stripePayment.OpctTxStatus = OpctTxInProgress
@@ -333,6 +348,7 @@ func Test_RetryIfTimedOut_Timed_Out(t *testing.T) {
 }
 
 func Test_DeleteStripePaymentIfExists(t *testing.T) {
+	t.Skip()
 	DeleteStripePaymentsForTest()
 	stripePayment, _ := returnValidStripePaymentForTest()
 
@@ -349,6 +365,7 @@ func Test_DeleteStripePaymentIfExists(t *testing.T) {
 }
 
 func Test_PurgeOldStripePayments(t *testing.T) {
+	t.Skip()
 	DeleteStripePaymentsForTest()
 	stripePaymentNew, _ := returnValidStripePaymentForTest()
 	DB.Create(&stripePaymentNew)
