+++ routes/routes_test.go
@@ -1,0 +1,17 @@
+package routes
+
+import (
+	"os"
+	"testing"
+)
+
+func TestMain(m *testing.M) {
+	os.Exit(testMain(m))
+}
+
+func testMain(m *testing.M) int {
+	setupRoutesTests()
+	cleanUpBeforeRoutesTest()
+
+	return m.Run()
+}
