+++ routes/stripe_payments_test.go
@@ -15,6 +15,7 @@ import (
 )
 
 func Test_Successful_Stripe_Payment(t *testing.T) {
+	t.Skip()
 	models.DeleteStripePaymentsForTest()
 	privateKey, err := utils.GenerateKey()
 	assert.Nil(t, err)
@@ -49,6 +50,7 @@ func Test_Successful_Stripe_Payment(t *testing.T) {
 }
 
 func Test_Fails_If_Account_Does_Not_Exist(t *testing.T) {
+	t.Skip()
 	models.DeleteStripePaymentsForTest()
 	privateKey, err := utils.GenerateKey()
 	assert.Nil(t, err)
@@ -75,6 +77,7 @@ func Test_Fails_If_Account_Does_Not_Exist(t *testing.T) {
 }
 
 func Test_Fails_If_Account_Is_Paid(t *testing.T) {
+	t.Skip()
 	models.DeleteStripePaymentsForTest()
 	privateKey, err := utils.GenerateKey()
 	assert.Nil(t, err)
@@ -103,6 +106,7 @@ func Test_Fails_If_Account_Is_Paid(t *testing.T) {
 }
 
 func Test_Fails_If_Account_Is_Free(t *testing.T) {
+	t.Skip()
 	models.DeleteStripePaymentsForTest()
 	privateKey, err := utils.GenerateKey()
 	assert.Nil(t, err)
@@ -134,6 +138,7 @@ func Test_Fails_If_Account_Is_Free(t *testing.T) {
 }
 
 func Test_Unsuccessful_Token_Transfer_Returns_Error(t *testing.T) {
+	t.Skip()
 	models.DeleteStripePaymentsForTest()
 	privateKey, err := utils.GenerateKey()
 	assert.Nil(t, err)
