+++ routes/upload_file.go
@@ -82,7 +82,7 @@ func uploadChunk(request UploadFileReq, c *gin.Context) error {
 	fileSize := len(request.ChunkData)
 	isLastChunk := request.uploadFileObj.PartIndex == file.EndIndex
 	if !isLastChunk && fileSize < int(utils.MinMultiPartSize) {
-		return BadRequestResponse(c, fmt.Errorf("Upload chunk is %v and does not meet min fileSize %v", fileSize, utils.MinMultiPartSize))
+		return BadRequestResponse(c, fmt.Errorf("upload chunk is %v and does not meet min fileSize %v", fileSize, utils.MinMultiPartSize))
 	}
 
 	completedPart, multipartErr := handleChunkData(file, request.uploadFileObj.PartIndex, []byte(request.ChunkData))
