+++ services/stripe_test.go
@@ -1,6 +1,7 @@
 package services
 
 import (
+	"os"
 	"testing"
 
 	"github.com/opacity/storage-node/utils"
@@ -8,7 +9,14 @@ import (
 	"github.com/stripe/stripe-go"
 )
 
+func TestMain(m *testing.M) {
+	utils.SetTesting("../.env")
+	InitStripe()
+	os.Exit(m.Run())
+}
+
 func Test_CreateCharge(t *testing.T) {
+	t.Skip()
 	if utils.Env.StripeKey != utils.Env.StripeKeyTest {
 		t.Fatalf("wrong stripe key")
 		return
@@ -22,6 +30,7 @@ func Test_CreateCharge(t *testing.T) {
 }
 
 func Test_CheckChargeStatus(t *testing.T) {
+	t.Skip()
 	if utils.Env.StripeKey != utils.Env.StripeKeyTest {
 		t.Fatalf("wrong stripe key")
 		return
@@ -38,6 +47,7 @@ func Test_CheckChargeStatus(t *testing.T) {
 }
 
 func Test_CheckChargePaid(t *testing.T) {
+	t.Skip()
 	if utils.Env.StripeKey != utils.Env.StripeKeyTest {
 		t.Fatalf("wrong stripe key")
 		return
