+++ services/stripe_test_utils.go
@@ -1,9 +1,6 @@
 package services
 
 import (
-	"os"
-	"testing"
-
 	"github.com/opacity/storage-node/utils"
 )
 
@@ -18,12 +15,6 @@ var testTokens = []string{
 	`tok_diners`,
 	`tok_jcb`,
 	`tok_unionpay`,
-}
-
-func TestMain(m *testing.M) {
-	utils.SetTesting("../.env")
-	InitStripe()
-	os.Exit(m.Run())
 }
 
 func RandTestStripeToken() string {
