+++ utils/utils_test.go
@@ -1,0 +1,11 @@
+package utils
+
+import (
+	"os"
+	"testing"
+)
+
+func TestMain(m *testing.M) {
+	SetTesting("../.env")
+	os.Exit(m.Run())
+}
